fix(integration): use secure tenant sentinels (#10895)
