+++ test/playwright/integration/healthcheck.spec.mjs
@@ -1,39 +1,8 @@
-import {test, expect}                   from '@playwright/test';
-import {createIdentityClient}           from './fixtures/identityClient.mjs';
+import {test, expect}                 from '@playwright/test';
+import {callHealthcheck, getReadiness} from './fixtures/mcpClient.mjs';
 
-const READY_URL = process.env.NEO_INTEGRATION_READY_URL || 'http://127.0.0.1:13090/ready';
-const KB_URL    = process.env.NEO_INTEGRATION_KB_URL    || 'http://127.0.0.1:13000';
-const MC_URL    = process.env.NEO_INTEGRATION_MC_URL    || 'http://127.0.0.1:13001';
-
-async function getReadiness() {
-    const response = await fetch(READY_URL);
-    return response.json();
-}
-
-async function callHealthcheck(baseUrl) {
-    const client = await createIdentityClient({
-        baseUrl,
-        clientName: 'neo-integration-healthcheck',
-        identity  : 'neo-healthcheck'
-    });
-
-    try {
-        const result = await client.callTool({name: 'healthcheck', arguments: {}});
-
-        expect(result.isError).not.toBe(true);
-
-        if (result.structuredContent) {
-            return result.structuredContent;
-        }
-
-        const text = result.content?.find(item => item.type === 'text')?.text;
-        expect(text, 'MCP healthcheck should return text content when structuredContent is absent').toBeTruthy();
-
-        return JSON.parse(text);
-    } finally {
-        await client.close();
-    }
-}
+const KB_URL = process.env.NEO_INTEGRATION_KB_URL || 'http://127.0.0.1:13000';
+const MC_URL = process.env.NEO_INTEGRATION_MC_URL || 'http://127.0.0.1:13001';
 
 test.describe('Dockerized KB/MC MCP healthcheck integration (#10805 Lane A)', () => {
     test('KB and MC expose healthcheck tool payloads over /mcp', async () => {
